pyproject.toml
# AI Image Generator - Project Configuration
# ==========================================

[project]
name = "ai-imageforge"
version = "1.0.0"
description = "Universal AI Image Generation with Automatic GPU Optimization"
authors = ["AI-ImageForge Team"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ai-imageforge"
repository = "https://github.com/ai-imageforge"
keywords = ["ai", "image-generation", "stable-diffusion", "gpu", "streamlit", "forge"]

[project.dependencies]
streamlit = ">=1.30.0"
Pillow = ">=10.4.0"
torch = ">=2.0.0"
torchvision = ">=0.15.0"
torchaudio = ">=2.0.0"
transformers = ">=4.30.0"
diffusers = ">=0.20.0"
accelerate = ">=0.20.0"
huggingface-hub = ">=0.16.0"
onnxruntime = ">=1.15.0"
requests = ">=2.28.0"
numpy = ">=1.24.0"
opencv-python = ">=4.8.0"
matplotlib = ">=3.6.0"
scipy = ">=1.10.0"
psutil = ">=5.9.0"

[project.optional-dependencies]
amd = ["onnxruntime-directml>=1.15.0"]
nvidia = ["torch[cuda]>=2.0.0"]
dev = ["pytest>=7.0.0", "black>=22.0.0", "flake8>=5.0.0"]

[tool.streamlit]
[tool.streamlit.server]
port = 8501
address = "localhost"
maxUploadSize = 200
maxMessageSize = 200

[tool.streamlit.theme]
primaryColor = "#FF6B6B"
backgroundColor = "#FFFFFF"
secondaryBackgroundColor = "#F0F2F6"
textColor = "#262730"

[tool.streamlit.browser]
gatherUsageStats = false

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
